extract sub components
